Исправил на long long

+++ print_free.c
@@ -3,7 +3,7 @@
 
 // печатает строку, фришит память + запоминает сколько байт вывел
 
-void		ft_print_free(char *str, t_lists *list, int max)
+void		ft_print_free(char *str, t_lists *list, long long max)
 {
 	int 	i;
